[servicedvb.cpp] fix showing now and next events in infobars config.usage.show_eit_nownext is now independent. Before this change the control was total of config.epg.eit. If you are using EPG IMPORT then dxNoEIT is also applied in the two information bars, therefore EIT is not shown in the information bars even if config.usage.show_eit_nownext was TRUE, the event information was always controlled by EPG IMPORT.
Now if both parameters are enabled, you can show EIT in the information bars, EIT takes control in both bars and the EPG cache from EPG import will be shown in other types of interface, simple EPG, graphics, channel selection, etc. .that is to say as it has always been.
Thanks to @norhap